Extract View Updates
